refactor: jwt 토큰 인증 필요 여부에 따라 public, secure로 api 분리 및 내부 애플리케이션끼리 통신하는 api는 internal로 분리
