+++ src/main/java/com/polar_moviechart/movieservice/domain/service/MovieRatingCommandService.java
@@ -1,6 +1,6 @@
 package com.polar_moviechart.movieservice.domain.service;
 
-import com.polar_moviechart.movieservice.domain.controller.UpdateRatingRequest;
+import com.polar_moviechart.movieservice.domain.controller.secureapi.UpdateRatingRequest;
 import com.polar_moviechart.movieservice.domain.entity.MovieRating;
 import com.polar_moviechart.movieservice.domain.repository.MovieRatingRepository;
 import lombok.RequiredArgsConstructor;
